Extract heading and description into traits
