+++ packages/infolists/src/Components/Section.php
@@ -3,7 +3,9 @@
 namespace Filament\Infolists\Components;
 
 use Closure;
+use Filament\Support\Concerns\HasDescription;
 use Filament\Support\Concerns\HasExtraAlpineAttributes;
+use Filament\Support\Concerns\HasHeading;
 use Filament\Support\Concerns\HasIcon;
 use Illuminate\Contracts\Support\Htmlable;
 use Illuminate\Support\Str;
@@ -13,17 +15,15 @@ class Section extends Component
     use Concerns\CanBeCollapsed;
     use Concerns\CanBeCompacted;
     use Concerns\EntanglesStateWithSingularRelationship;
+    use HasDescription;
     use HasExtraAlpineAttributes;
+    use HasHeading;
     use HasIcon;
 
     /**
      * @var view-string
      */
     protected string $view = 'filament-infolists::components.section';
-
-    protected string | Htmlable | Closure | null $description = null;
-
-    protected string | Htmlable | Closure $heading;
 
     protected bool | Closure | null $isAside = null;
 
@@ -49,35 +49,11 @@ class Section extends Component
         $this->columnSpan('full');
     }
 
-    public function description(string | Htmlable | Closure | null $description = null): static
-    {
-        $this->description = $description;
-
-        return $this;
-    }
-
-    public function heading(string | Htmlable | Closure $heading): static
-    {
-        $this->heading = $heading;
-
-        return $this;
-    }
-
     public function aside(bool | Closure | null $condition = true): static
     {
         $this->isAside = $condition;
 
         return $this;
-    }
-
-    public function getDescription(): string | Htmlable | null
-    {
-        return $this->evaluate($this->description);
-    }
-
-    public function getHeading(): string | Htmlable
-    {
-        return $this->evaluate($this->heading);
     }
 
     public function getId(): ?string
